+++ Source/WebCore/bindings/js/JSDOMMapLike.cpp
@@ -29,7 +29,6 @@
 #include "WebCoreJSBuiltinInternals.h"
 #include "WebCoreJSClientData.h"
 #include <JavaScriptCore/CatchScope.h>
-#include <JavaScriptCore/HashMapImplInlines.h>
 #include <JavaScriptCore/JSMap.h>
 #include <JavaScriptCore/VMTrapsInlines.h>
